Renamed custom_simulator to simulator_custom for consistency.
